refactor: 에러 핸들링 수정
