+++ app.js
@@ -15,10 +15,12 @@ connectMongoDB();
 
 const app = express();
 
-app.use(cors({
-  origin: true,
-  credentials: true
-}));
+app.use(
+  cors({
+    origin: true,
+    credentials: true,
+  })
+);
 
 app.use(logger("dev"));
 app.use(express.json());
