+++ middlewares/catchError.js
@@ -1,12 +1,12 @@
 const ERROR = require("../constants/error");
 
-exports.catchError = (fn, errorType, errorStatus) => async (req, res, next) => {
+exports.catchError = (fn) => async (req, res, next) => {
   try {
     await fn(req, res, next);
   } catch (err) {
-    console.log(err);
-    const error = new Error(ERROR[errorType]);
-    error.status = errorStatus;
-    next(error);
+    err.message = err.message || ERROR.SERVER_ERROR;
+    err.status = err.message ? 400 : 500;
+
+    next(err);
   }
 };
