+++ routes/controllers/otpController.js
@@ -1,7 +1,7 @@
 const User = require("../../models/User");
 
 const sendEmail = require("../../utils/sendEmail");
-const createLowerUpperNumPassword = require("../../services/createLowerUpperNumPassword");
+const { createLowerUpperNumPassword } = require("../../services/createRandomPasswords");
 
 const ERROR = require("../../constants/error");
 
@@ -10,30 +10,19 @@ module.exports = {
     const { email } = req.body;
     const password = createLowerUpperNumPassword(12);
 
-    try {
-      const user = await User.findOneAndUpdate(
-        { email },
-        { $set: { oneTimePassword: password } }
-      );
+    const user = await User.findOneAndUpdate(
+      { email },
+      { $set: { oneTimePassword: password } }
+    );
 
-      if (!user) {
-        throw new Error(ERROR.NO_ACCOUNT);
-      }
-    } catch (err) {
-      err.status = 400;
-      return next(err);
+    if (!user) {
+      throw new Error(ERROR.NO_ACCOUNT);
     }
 
-    try {
-      const emailTitle = "One Time Password";
-      const emailBody = `<h2>Hello OnePass User!</h2><br /><h3>This OTP is valid for only once. Please Login with this OTP and securely change your password in 'My Account'.</h3><h2>OTP: ${password}</h2><br /><h3>Best Regards,</h3><h3>One Pass Team</h3>`;
+    const emailTitle = "One Time Password";
+    const emailBody = `<h2>Hello OnePass User!</h2><br /><h3>This OTP is valid for only once. Please Login with this OTP and securely change your password in 'My Account'.</h3><h2>OTP: ${password}</h2><br /><h3>Best Regards,</h3><h3>One Pass Team</h3>`;
 
-      await sendEmail(email, emailTitle, emailBody);
-      res.status(200).json({ data: password });
-    } catch (err) {
-      const error = new Error(ERROR.SERVER_ERROR);
-      error.status = 500;
-      next(error);
-    }
+    await sendEmail(email, emailTitle, emailBody);
+    res.status(200).json({ data: password });
   },
 };
