+++ routes/login.js
@@ -1,6 +1,8 @@
 const express = require("express");
 const router = express.Router();
 
+const { catchError } = require("../middlewares/catchError");
+
 const {
   login,
   sendVerifier,
@@ -8,13 +10,11 @@ const {
   deleteOTP,
 } = require("./controllers/loginController");
 
-const { catchError } = require("../middlewares/catchError");
-
-router.post("/", catchError(login, "FAIL_LOGIN", 400));
-router.get("/:email", catchError(sendVerifier, "NO_ACCOUNT", 400));
+router.post("/", login);
+router.get("/:email", sendVerifier);
 router
   .route("/otp/:email")
-  .get(catchError(checkOTP, "SERVER_ERROR", 500))
-  .delete(catchError(deleteOTP, "SERVER_ERROR", 500));
+  .get(catchError(checkOTP))
+  .delete(catchError(deleteOTP));
 
 module.exports = router;
