+++ routes/logout.js
@@ -2,9 +2,8 @@ const express = require("express");
 const router = express.Router();
 
 const { logout } = require("./controllers/logoutController");
-
 const { catchError } = require("../middlewares/catchError");
 
-router.post("/:userId", catchError(logout, "SERVER_ERROR", 500));
+router.post("/:userId", catchError(logout));
 
 module.exports = router;
