+++ routes/otp.js
@@ -2,7 +2,8 @@ const express = require("express");
 const router = express.Router();
 
 const { sendOTP } = require("./controllers/otpController");
+const { catchError } = require("../middlewares/catchError");
 
-router.post("/", sendOTP);
+router.post("/", catchError(sendOTP));
 
 module.exports = router;
