+++ routes/users.js
@@ -2,6 +2,7 @@ const express = require("express");
 const router = express.Router();
 
 const ensureAuthenticated = require("../middlewares/ensureAuthenticated");
+const { catchError } = require("../middlewares/catchError");
 
 const {
   getUserInfo,
@@ -17,54 +18,41 @@ const {
   createRandomPassword,
 } = require("./controllers/usersController");
 
-const { catchError } = require("../middlewares/catchError");
-
-router.get(
-  "/:userId",
-  ensureAuthenticated,
-  catchError(getUserInfo, "SERVER_ERROR", 500)
-);
+router.get("/:userId", ensureAuthenticated, catchError(getUserInfo));
 
 router
   .route("/:userId/password")
-  .post(ensureAuthenticated, catchError(addPassword, "SERVER_ERROR", 500));
+  .post(ensureAuthenticated, catchError(addPassword));
 
 router
   .route("/:userId/password/:passwordId")
-  .get(ensureAuthenticated, catchError(getPassword, "SERVER_ERROR", 500))
-  .patch(ensureAuthenticated, catchError(updatePassword, "SERVER_ERROR", 500))
-  .delete(ensureAuthenticated, catchError(deletePassword, "SERVER_ERROR", 500));
+  .get(ensureAuthenticated, catchError(getPassword))
+  .patch(ensureAuthenticated, catchError(updatePassword))
+  .delete(ensureAuthenticated, catchError(deletePassword));
 
 router.delete(
   "/:userId/withdraw",
   ensureAuthenticated,
-  catchError(deleteAccount, "SERVER_ERROR", 500)
+  catchError(deleteAccount)
 );
 
 router.patch(
   "/:userId/reset-password",
   ensureAuthenticated,
-  catchError(changeMasterPassword, "SERVER_ERROR", 500)
+  catchError(changeMasterPassword)
 );
 
 router
   .route("/:userId/account-setting")
-  .get(ensureAuthenticated, catchError(getAccountSetting, "SERVER_ERROR", 500))
-  .post(
-    ensureAuthenticated,
-    catchError(setAccountSetting, "SERVER_ERROR", 500)
-  );
+  .get(ensureAuthenticated, catchError(getAccountSetting))
+  .post(ensureAuthenticated, catchError(setAccountSetting));
 
-router.get(
-  "/:userId/url/:url",
-  ensureAuthenticated,
-  catchError(checkUrlData, "SERVER_ERROR", 500)
-);
+router.get("/:userId/url/:url", ensureAuthenticated, catchError(checkUrlData));
 
 router.get(
   "/:userId/random-password",
   ensureAuthenticated,
-  catchError(createRandomPassword, "SERVER_ERROR", 500)
+  catchError(createRandomPassword)
 );
 
 module.exports = router;
