--- services/createLowerUpperNumSymbolPassword.js
@@ -1,16 +1,0 @@
-function createLowerUpperNumSymbolPassword(length) {
-  let randomString = "";
-  const characters =
-    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789~!@#$%^&*_-+=<>?";
-  const charactersLength = characters.length;
-
-  for (let i = 0; i < length; i++) {
-    randomString += characters.charAt(
-      Math.floor(Math.random() * charactersLength)
-    );
-  }
-
-  return randomString;
-}
-
-module.exports = createLowerUpperNumSymbolPassword;
